pyproject.toml
[project]
name = "ft-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp[speedups]>=3.13.5",
    "aiosqlite>=0.22.1",
    "argon2-cffi>=25.1.0",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.135.3",
    "python-dotenv>=1.2.2",
    "python-ms>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "sqlmodel>=0.0.38",
]


[tool.ruff]
line-length = 90
target-version = "py314"

[tool.ruff.lint]
extend-select = ["I"]  # Enables import sorting

[tool.ruff.lint.isort]
known-first-party = ["commands", "events", "lib", "logic", "models"]
combine-as-imports = true
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

